refactor(gamma): extract member-management orchestration out of GroupDetailPage
Aikido re-flagged GroupDetailPage.tsx as too long — it grew to 576
lines after this PR added five member-management handlers on top of
the extraction already done for FOUND-103, exceeding the 359-line
version that got flagged in the first place.

Moved the add/invite/edit/remove-member and delete-invitation
workflows (their sheet-open state, mutations, and notifications) into
a new useGroupMemberActions hook, matching Aikido's own suggested fix
("extract member lifecycle logic into a hook/service"). The page
component now only wires the hook's returned state and handlers into
its JSX.

Pure refactor, no behavior change — verified by the existing
GroupDetailPage.spec.tsx suite passing unchanged (it mocks the
underlying useGroupMutations/useGroupDetail hooks the new hook calls
internally, so it exercises the extracted logic for real without
needing any test changes).
